+++ app/TestModule/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+TestModule = Blueprint("TestModule", __name__)
+
+from app.TestModule import routes
